Fix compatibility with Ruby 1.9.3 and 2.0.0
The named arguments syntax used for the Rails 5 `digest` method is only
valid on Ruby 2.1 or higher (or 2.0.0 if default values are supplied),
causing an error parsing the lib/rabl/digestor.rb source on these
older versions.

    SyntaxError: rabl-0.13.0/lib/rabl/digestor.rb:6: syntax error, unexpected ','
          def self.digest(name:, finder:, **options)

The support for different Rails versions has been moved into separate
files that can be loaded only when required.
