+++ .travis.yml
@@ -5,7 +5,7 @@ rvm:
   - 1.9.3
   - 2.0.0
   - 2.1
-  - 2.2
+  - 2.2.5
   - 2.3.1
 notifications:
   recipients:
