+++ Gemfile
@@ -14,9 +14,10 @@ group :test do
   # RABL TEST
   gem 'builder'
 
+  rails_version = RUBY_VERSION =~ /\A(1|2.[01])/ ? '~> 4.0' : '>= 4.0'
   # FIXTURES
   gem 'rack-test', :require => 'rack/test'
-  gem 'activerecord', :require => 'active_record'
+  gem 'activerecord', rails_version, :require => 'active_record'
   gem 'sqlite3'
   gem 'sinatra', '>= 1.2.0'
   gem 'hashie'
